requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
pymongo==4.6.0
python-dotenv==1.0.0
certifi==2024.8.30
pyOpenSSL==24.2.1